+++ packages/data-schema/src/ModelType.ts
@@ -7,6 +7,7 @@ import { ModelField, InternalField } from './ModelField';
 import type {
   ModelRelationalField,
   InternalRelationalField,
+  ModelRelationalFieldParamShape,
 } from './ModelRelationalField';
 import { Authorization } from './Authorization';
 import { RefType } from './RefType';
@@ -131,6 +132,27 @@ type ConflictingAuthRulesMap<T extends ModelTypeParamShape> = {
     : never;
 };
 
+export type AddRelationshipFieldsToModelTypeFields<
+  Model,
+  RelationshipFields extends Record<
+    string,
+    ModelRelationalField<ModelRelationalFieldParamShape, string, any, any>
+  >,
+> =
+  Model extends ModelType<
+    infer ModelParam extends ModelTypeParamShape,
+    infer HiddenKeys
+  >
+    ? ModelType<
+        SetTypeSubArg<
+          ModelParam,
+          'fields',
+          ModelParam['fields'] & RelationshipFields
+        >,
+        HiddenKeys
+      >
+    : never;
+
 /**
  * For a given ModelTypeParamShape, produces a union of Authorization rules
  * that would *conflict* with the given type.
@@ -211,18 +233,37 @@ export type ModelType<
  * Used on the complete schema object.
  */
 export type SchemaModelType<
-  T extends ModelType<ModelTypeParamShape> = ModelType<ModelTypeParamShape>,
-> = T & {
-  addRelationships(
-    relationships: Record<string, ModelRelationalField<any, string, any, any>>,
-  ): void;
-};
+  T extends ModelType<ModelTypeParamShape, never | 'identifier'> = ModelType<
+    ModelTypeParamShape,
+    never | 'identifier'
+  >,
+  ModelName extends string = string,
+  IsRDS extends boolean = false,
+> = IsRDS extends true
+  ? T & {
+      addRelationships<
+        Param extends Record<
+          string,
+          ModelRelationalField<any, string, any, any>
+        > = Record<never, never>,
+      >(
+        relationships: Param,
+      ): Record<ModelName, Param>;
+      fields: T extends ModelType<infer R extends ModelTypeParamShape, any>
+        ? R['fields']
+        : never;
+    }
+  : T;
 
 /**
  * Internal representation of Model Type that exposes the `data` property.
  * Used at buildtime.
  */
-export type InternalModel = SchemaModelType & {
+export type InternalModel = SchemaModelType<
+  ModelType<ModelTypeParamShape>,
+  string,
+  true
+> & {
   data: InternalModelData;
 };
 
@@ -259,6 +300,7 @@ function _model<T extends ModelTypeParamShape>(fields: T['fields']) {
     addRelationships(relationships) {
       data.fields = { ...data.fields, ...relationships };
     },
+    fields: data.fields,
   } as InternalModel as ModelType<T>;
 }
 
@@ -299,6 +341,3 @@ export function model<T extends ModelFields>(
 }> {
   return _model(fields);
 }
-
-// TODO: rename and extract into separate file;
-// Will breaking apart SecondaryIndexToIR optimize it?