+++ packages/data-schema/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/data-schema
 
+## 0.14.11
+
+### Patch Changes
+
+- 14a1951: fix: Change interface to move schema combination processing and expose rds config
+- Updated dependencies [14a1951]
+  - @aws-amplify/data-schema-types@0.7.16
+
+## 0.14.10
+
+### Patch Changes
+
+- 9dfc6a9: feat: add .relationships() modifiers for RDSSchema
+
 ## 0.14.9
 
 ### Patch Changes